client/client_requirements.txt
pvleopard
requests
rpi-lgpio
RPi.GPIO
gpiozero
python-dotenv
pvrecorder
pvporcupine
mutagen
psutil
